ai-cad-backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pydantic
python-multipart
